+++ packages/assets-controllers/src/crypto-compare-service/crypto-compare.ts
@@ -38,29 +38,30 @@ function getPricingURL(
 
 /**
  * Get the CryptoCompare API URL for getting the conversion rate from a given array of native currencies
- * to the given currency. Optionally, the conversion rate from the native currency to USD can also be
+ * to the given currencies. Optionally, the conversion rate from the native currency to USD can also be
  * included in the response.
  *
  * @param fsyms - The native currencies to get conversion rates for.
- * @param tsyms - The currency to convert to.
+ * @param tsyms - The currencies to convert to.
  * @param includeUSDRate - Whether or not the native currency to USD conversion rate should be included.
  * @returns The API URL for getting the conversion rates.
  */
 function getMultiPricingURL(
-  fsyms: string,
-  tsyms: string,
+  fsyms: string[],
+  tsyms: string[],
   includeUSDRate = false,
 ) {
   const updatedTsyms =
-    includeUSDRate && !tsyms.includes('USD') ? `${tsyms},USD` : tsyms;
+    includeUSDRate && !tsyms.some((t) => t.toUpperCase() === 'USD')
+      ? [...tsyms, 'USD']
+      : tsyms;
 
   const params = new URLSearchParams();
-  params.append('fsyms', fsyms);
-  params.append('tsyms', updatedTsyms);
+  params.append('fsyms', fsyms.join(','));
+  params.append('tsyms', updatedTsyms.join(','));
 
   const url = new URL(`${CRYPTO_COMPARE_DOMAIN}/data/pricemulti`);
   url.search = params.toString();
-
   return url.toString();
 }
 
@@ -139,19 +140,19 @@ export async function fetchMultiExchangeRate(
   fiatCurrency: string,
   cryptocurrencies: string[],
   includeUSDRate: boolean,
-): Promise<Record<string, Record<string, string>>> {
+): Promise<Record<string, Record<string, number>>> {
   const url = getMultiPricingURL(
-    Object.values(cryptocurrencies).join(','),
-    fiatCurrency,
+    cryptocurrencies,
+    [fiatCurrency],
     includeUSDRate,
   );
   const response = await handleFetch(url);
   handleErrorResponse(response);
 
-  const rates: Record<string, Record<string, string>> = {};
+  const rates: Record<string, Record<string, number>> = {};
   for (const [cryptocurrency, values] of Object.entries(response) as [
     string,
-    Record<string, string>,
+    Record<string, number>,
   ][]) {
     rates[cryptocurrency.toLowerCase()] = {
       [fiatCurrency.toLowerCase()]: values[fiatCurrency.toUpperCase()],